tweet_collector/requirements.txt
tweepy
pandas
sqlalchemy
python-dotenv
psycopg2-binary
vaderSentiment